chore(benches): decompose kitchen-sink bench.rs into focused files (#426)
Split benches/bench.rs (454 lines, 6 functions across analysis / vector
/ spell / parallel / memory / synonym) into three focused files and drop
the rest.

New files (all conform to the #427 hygiene policy):
- benches/text_analysis_bench.rs
- benches/spell_correction_bench.rs
- benches/synonym_bench.rs

Functions retained:
- bench_text_analysis     -> text_analysis_bench.rs
- bench_synonym_dictionary -> synonym_bench.rs
- bench_spell_correction  -> spell_correction_bench.rs (rewritten to use
  iter_batched with a fresh SpellingCorrector per iteration; the previous
  shared `&mut corrector` mutated query_history across iterations and
  conflated cold and warm-cache paths)

Functions dropped:
- bench_vector_search / vector_operations_basic — duplicates of
  vector_search_bench.rs / distance_bench.rs; vector_operations_basic
  also hand-rolled a top-50 search loop, misleading as a signal.
- bench_memory_operations — Vec::push cost is not on any hot path.
- bench_parallel_operations — measured rayon overhead, unrelated to any
  audited hot path.
- bench_scalability — duplicate of bench_vector_search at larger scale.

Cargo.toml: drop the bench entry, register the three new benches.

Net change: ~ -100 lines, clearer ownership per file.

Closes #426
